Refactor tRPC client code
Start a new `@/client` path to complement `@/server`
