--- projects/app/src/util/trpc.ts
@@ -1,4 +1,0 @@
-import type { AppRouter } from "@/server/routers/_app";
-import { createTRPCReact } from "@trpc/react-query";
-
-export const trpc = createTRPCReact<AppRouter>();
